refactor : 공용 UI 잔여 로컬 컴포넌트 파일 분리
* 알림 팝오버의 로딩·에러 폴백을 형제 파일로 분리
* 평점 표시·입력에 중복된 별 SVG를 RatingStar 아이콘으로 통합
* 날짜 입력의 달력 SVG를 Calendar 아이콘으로 분리
