+++ apps/web/src/shared/ui/icons/Calendar.tsx
@@ -1,0 +1,20 @@
+interface CalendarProps {
+  className?: string;
+}
+
+/** 달력. 크기·색은 className으로(currentColor). */
+export function Calendar({ className = "text-[1.125rem] text-ink-3" }: CalendarProps) {
+  return (
+    <svg
+      className={className}
+      width="1em"
+      height="1em"
+      viewBox="0 0 24 24"
+      fill="none"
+      aria-hidden
+    >
+      <rect x="3" y="4.5" width="18" height="16" rx="2.5" stroke="currentColor" strokeWidth="1.6" />
+      <path d="M3 9h18M8 2.5v4M16 2.5v4" stroke="currentColor" strokeWidth="1.6" strokeLinecap="round" />
+    </svg>
+  );
+}
